+++ server/backend/src/main/java/com/cropify/services/impl/FarmProductsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.cropify.services;
+package com.cropify.services.impl;
 
 import java.util.List;
 import java.util.Optional;
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Service;
 import com.cropify.customexception.ResourceNotFoundException;
 import com.cropify.dao.FarmProductsRepository;
 import com.cropify.entity.FarmProducts;
+import com.cropify.services.FarmProductsService;
 
 @Service
 @Transactional
